requirements.txt
dataclasses-json==0.6.4
firebase-admin==6.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
openai==1.30.5
python-dotenv==1.0.1
streamlit==1.31.0
toml==0.10.2
authlib==1.3.0
requests==2.31.0
# flask==3.0.3

